Updated Objects to all be children of one class to help with managing them
Added InteractableObject Parent Actor,
Made InteractableIngredient and InteractableSpawner, which have different attributes and can be interacted with by the player.
Updated Logic for Player when Interacting with an Object.
Added Enums for different types of Object and Ingredient.
Added an Identifier, which checks what type of ingredient a thing is, can be used later on.
IMO simplified Logic.